python-Agent/pyproject.toml
[project]
name = "python-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "vision-agents[deepgram,elevenlabs,gemini,getstream]>=0.4.7",
    "fastapi",
    "uvicorn",
    "python-dotenv",
]
